The new QR Hiding2
Lightmaps still need to be baked fml this takes long,
Assets nicely play with fore and background, screen is filled but not cluttered, added lights to the scene to emphasize atmosphere of parts of level, also my postprocessing stack is at 2x saturation, so now all of yours are too. P.s. also added to the camzone script to enable having orthographic cameras!